Patch by Ben Gras fixing a few minor issues: adds missing includes, fixes the ntohs prototype in klee-libc, and removes some unused code.
git-svn-id: https://llvm.org/svn/llvm-project/klee/trunk@146352 91177308-0d34-0410-b5e6-96231b3b80d8

+++ runtime/klee-libc/htonl.c
@@ -41,7 +41,7 @@ uint32_t htonl(uint32_t v) {
 
 #endif
 
-uint16_t ntohs(uint32_t v) {
+uint16_t ntohs(uint16_t v) {
   return htons(v);
 }
 uint32_t ntohl(uint32_t v) {
